Dialog!
Added a function to handle dialog. Upon interacting with a "talkable"
tagged object, the game will search for the xml file that contains the
message to be displayed and shows it. The dialog box resizes dynamically
(though I will have it initialize at a size and stay that way for better
framerate) and the text adds newline characters as needed to fix the
text.