Mobile Controls
We now have mobile controls! There is a joystick (Thanks to Ferret
Syndicate) in the left corner, and an A and B button on the right side.
There is also a pause button on the top of the screen. The B button and
pause button do not yet have a function. TODO: Pause function and Scene
switching (entering and exiting buildings)