Changes to saving
Deleted a few files, but in the previous commit I forgot to mention that
I revamped the saving system, and eliminated XML completely. The game
now used PlayerPreferences, which persist after the game is destroyed.
Saving will be added soon. Dialog also now works on all systems, as
there is now a class dedicated to text. This should not only be faster,
but it should make the game almost completely universal.